change the type of Leaderboard::Player::score from "int" to "long long"

+++ src/leaderboard.cpp
@@ -2,15 +2,15 @@
 
 Leaderboard::Player::Player() {};
 
-Leaderboard::Player::Player(const time_t &time, const int &score) {
+Leaderboard::Player::Player(const time_t &time, const long long &score) {
   this->time = time;
   this->score = score;
 }
 
 Leaderboard::Leaderboard(FILE *const &f) {
   time_t time;
-  int score;
-  while(~fscanf(f, "%ld%d", &time, &score)) {
+  long long score;
+  while(~fscanf(f, "%ld%lld", &time, &score)) {
     insert(time, score);
   }
   fclose(f);
@@ -20,7 +20,7 @@ void Leaderboard::reset() {
   players.clear();
 }
 
-void Leaderboard::insert(const time_t &time, const int &score) {
+void Leaderboard::insert(const time_t &time, const long long &score) {
   players.emplace_back(time, score);
   std::sort(players.begin(), players.end(),
             [](const auto &p1, const auto &p2) -> bool {
@@ -36,7 +36,7 @@ void Leaderboard::insert(const time_t &time, const int &score) {
 void Leaderboard::save() const {
   auto f = fopen(LEADERBOARD_PATH, "w+");
   for(const auto &[time, score]:players) {
-    fprintf(f, "%ld %d", time, score);
+    fprintf(f, "%ld %lld", time, score);
   }
   fclose(f);
 }
